🐝 move Post types to a separate file
